split out api, test_utils, and bundlerepo from cli
