+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["cli", "codeowners", "junit-mock", "context"]
+members = ["api", "cli", "codeowners", "context", "junit-mock", "test_utils"]
 resolver = "2"
 
 [profile.release]
