+++ cli/Cargo.toml
@@ -13,8 +13,10 @@ path = "src/lib.rs"
 
 [dependencies]
 anyhow = "1.0.44"
+api = { path = "../api" }
 chrono = { version = "0.4.33", default-features = false, features = ["clock"] }
 clap = { version = "4.4.18", features = ["derive", "env"] }
+context = { path = "../context" }
 env_logger = { version = "0.11.0", default-features = false }
 log = "0.4.14"
 exitcode = "1.1.1"
@@ -24,7 +26,6 @@ tokio = { version = "*", default-features = false, features = [
 ] }
 tempfile = "3.2.0"
 tokio-retry = { version = "0.3", default-features = false }
-gix = { version = "0.63.0", default-features = false, features = [] }
 glob = "0.3.0"
 regex = { version = "1.10.3", default-features = false, features = ["std"] }
 reqwest = { version = "0.12.5", default-features = false, features = [
@@ -46,8 +47,8 @@ uuid = { version = "1.10.0", features = ["v5"] }
 assert_cmd = "2.0.16"
 assert_matches = "1.5.0"
 axum = { version = "0.7.5", features = ["macros"] }
-git2 = "0.19.0"                                     # Used for creating test repos with libgit2
 junit-mock = { path = "../junit-mock" }
+test_utils = { path = "../test_utils" }
 
 [build-dependencies]
 vergen = { version = "8.3.1", features = [
