+++ cli/src/scanner.rs
@@ -7,16 +7,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::codeowners::{CodeOwners, Owners};
 use crate::constants::{ALLOW_LIST, ENVS_TO_GET};
-use crate::types::{BundledFile, FileSetType, Repo};
-use crate::utils::from_non_empty_or_default;
-
-pub const GIT_REMOTE_ORIGIN_URL_CONFIG: &str = "remote.origin.url";
-
-#[derive(Debug)]
-struct HeadAuthor {
-    pub name: String,
-    pub email: String,
-}
+use crate::types::{BundledFile, FileSetType};
 
 #[derive(Default, Debug)]
 pub struct FileSetCounter {
@@ -142,146 +133,6 @@ impl FileSet {
             files,
             glob: glob_path,
         })
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct BundleRepo {
-    pub repo: Repo,
-    pub repo_root: String,
-    pub repo_url: String,
-    pub repo_head_sha: String,
-    pub repo_head_branch: String,
-    pub repo_head_commit_epoch: i64,
-    pub repo_head_commit_message: String,
-    pub repo_head_author_name: String,
-    pub repo_head_author_email: String,
-}
-
-impl BundleRepo {
-    /// Read important fields from git repo root.
-    ///
-    pub fn try_read_from_root(
-        in_repo_root: Option<String>,
-        in_repo_url: Option<String>,
-        in_repo_head_sha: Option<String>,
-        in_repo_head_branch: Option<String>,
-        in_repo_head_commit_epoch: Option<String>,
-    ) -> anyhow::Result<BundleRepo> {
-        let mut out_repo_url = in_repo_url.clone();
-        let mut out_repo_head_sha = in_repo_head_sha.clone();
-        let mut out_repo_head_branch = in_repo_head_branch.clone();
-        let mut out_repo_head_commit_epoch =
-            from_non_empty_or_default(in_repo_head_commit_epoch, None, |s| {
-                Some(
-                    s.parse::<i64>()
-                        .expect("failed to parse commit epoch override"),
-                )
-            });
-        let out_repo_root =
-            from_non_empty_or_default(in_repo_root, Self::default_to_working_directory(), Some);
-
-        let mut git_head_author = None;
-        let mut git_head_commit_message = None;
-        // If repo root found, try to get repo details from git.
-        if let Some(repo_root) = &out_repo_root {
-            // Read git repo.
-            log::info!("Reading git repo at {:?}", &repo_root);
-
-            let git_repo = gix::open(repo_root)?;
-            let git_url = git_repo
-                .config_snapshot()
-                .string_by_key(GIT_REMOTE_ORIGIN_URL_CONFIG)
-                .map(|s| s.to_string());
-            let mut git_head = git_repo.head()?;
-
-            let mut git_head_branch = git_head.referent_name().map(|s| s.as_bstr().to_string());
-            if git_head_branch.is_none() {
-                for r in git_repo.references()?.remote_branches()? {
-                    match r {
-                        Ok(r) => {
-                            let target = r.target();
-                            let id = target.try_id();
-                            if id.is_some()
-                                && git_head.id().is_some()
-                                && id.unwrap().to_string() == git_head.id().unwrap().to_string()
-                            {
-                                git_head_branch =
-                                    r.name().to_path().to_str().map(|s| s.to_string());
-                                break;
-                            };
-                        }
-                        Err(e) => {
-                            log::debug!("Unexpected error when trying to find reference {:?}", e);
-                        }
-                    }
-                }
-            }
-            let git_head_sha = git_head.id().map(|id| id.to_string());
-            let git_head_commit_time = git_head.peel_to_commit_in_place()?.time()?;
-            git_head_commit_message = git_head.peel_to_commit_in_place().map_or(None, |commit| {
-                commit
-                    .message()
-                    .map_or(None, |msg| Some(msg.title.to_string()))
-            });
-            git_head_author = git_head
-                .peel_to_commit_in_place()
-                .map(|commit| {
-                    if let Ok(author) = commit.author() {
-                        Some(HeadAuthor {
-                            name: author.name.to_string(),
-                            email: author.email.to_string(),
-                        })
-                    } else {
-                        None
-                    }
-                })
-                .ok()
-                .flatten();
-            log::info!("Found git_url: {:?}", git_url);
-            log::info!("Found git_sha: {:?}", git_head_sha);
-            log::info!("Found git_branch: {:?}", git_head_branch);
-            log::info!("Found git_commit_time: {:?}", git_head_commit_time);
-            log::info!("Found git_commit_message: {:?}", git_head_commit_message);
-            log::info!("Found git_author: {:?}", git_head_author);
-
-            out_repo_url = from_non_empty_or_default(in_repo_url, git_url, Some);
-            out_repo_head_sha = from_non_empty_or_default(in_repo_head_sha, git_head_sha, Some);
-            out_repo_head_branch =
-                from_non_empty_or_default(in_repo_head_branch, git_head_branch, Some);
-            if out_repo_head_commit_epoch.is_none() {
-                out_repo_head_commit_epoch = Some(git_head_commit_time.seconds);
-            }
-        }
-
-        // Require URL which should be known at this point.
-        let repo_url = out_repo_url.expect("failed to get repo url");
-        let repo = Repo::from_url(&repo_url)?;
-        let (git_head_author_name, git_head_author_email) = if let Some(author) = git_head_author {
-            (author.name, author.email)
-        } else {
-            (String::default(), String::default())
-        };
-
-        Ok(BundleRepo {
-            repo,
-            repo_root: out_repo_root.unwrap_or("".to_string()),
-            repo_url,
-            repo_head_branch: out_repo_head_branch.unwrap_or_default(),
-            repo_head_sha: out_repo_head_sha.expect("failed to get repo head sha"),
-            repo_head_commit_epoch: out_repo_head_commit_epoch
-                .expect("failed to get repo head commit time"),
-            repo_head_commit_message: git_head_commit_message.unwrap_or("".to_string()),
-            repo_head_author_name: git_head_author_name,
-            repo_head_author_email: git_head_author_email,
-        })
-    }
-
-    fn default_to_working_directory() -> Option<String> {
-        std::env::current_dir()
-            .expect("failed to resolve working directory")
-            .to_str()
-            .map(|s| s.to_string())
     }
 }
