+++ cli/tests/upload.rs
@@ -2,19 +2,18 @@ use std::fs;
 use std::io::BufReader;
 use std::path::Path;
 
+use api::{CreateBundleUploadRequest, CreateRepoRequest, GetQuarantineBulkTestStatusRequest};
 use assert_cmd::Command;
 use assert_matches::assert_matches;
+use context::repo::RepoUrlParts;
 use junit_mock::JunitMock;
 use tempfile::tempdir;
-use test_utils::mock_git_repo::setup_repo_with_commit;
-use test_utils::mock_server::{spawn_mock_server, RequestPayload};
-use trunk_analytics_cli::codeowners::CodeOwners;
-use trunk_analytics_cli::types::{
-    BundleMeta, CreateBundleUploadRequest, CreateRepoRequest, FileSetType,
-    GetQuarantineBulkTestStatusRequest, Repo,
+use test_utils::{
+    mock_git_repo::setup_repo_with_commit,
+    mock_server::{spawn_mock_server, RequestPayload},
 };
-
-mod test_utils;
+use trunk_analytics_cli::codeowners::CodeOwners;
+use trunk_analytics_cli::types::{BundleMeta, FileSetType};
 
 fn generate_mock_git_repo<T: AsRef<Path>>(directory: T) {
     setup_repo_with_commit(directory).unwrap();
@@ -71,7 +70,7 @@ async fn upload_bundle() {
     assert_eq!(
         requests_iter.next().unwrap(),
         RequestPayload::GetQuarantineBulkTestStatus(GetQuarantineBulkTestStatusRequest {
-            repo: Repo {
+            repo: RepoUrlParts {
                 host: String::from("github.com"),
                 owner: String::from("trunk-io"),
                 name: String::from("analytics-cli"),
@@ -83,7 +82,7 @@ async fn upload_bundle() {
     assert_eq!(
         requests_iter.next().unwrap(),
         RequestPayload::CreateBundleUpload(CreateBundleUploadRequest {
-            repo: Repo {
+            repo: RepoUrlParts {
                 host: String::from("github.com"),
                 owner: String::from("trunk-io"),
                 name: String::from("analytics-cli"),
@@ -102,7 +101,7 @@ async fn upload_bundle() {
     assert_eq!(bundle_meta.org, "test-org");
     assert_eq!(
         bundle_meta.repo.repo,
-        Repo {
+        RepoUrlParts {
             host: String::from("github.com"),
             owner: String::from("trunk-io"),
             name: String::from("analytics-cli"),
@@ -159,7 +158,7 @@ async fn upload_bundle() {
     assert_eq!(
         requests_iter.next().unwrap(),
         RequestPayload::CreateRepo(CreateRepoRequest {
-            repo: Repo {
+            repo: RepoUrlParts {
                 host: String::from("github.com"),
                 owner: String::from("trunk-io"),
                 name: String::from("analytics-cli"),
