+++ test_utils/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+api = { path = "../api" }
+anyhow = "1.0.44"
+axum = { version = "0.7.5", features = ["macros"] }
+git2 = "0.19.0"
+junit-mock = { path = "../junit-mock" }
+tar = { version = "0.4.30", default-features = false }
+tempfile = "3.2.0"
+tokio = { version = "*", default-features = false, features = [
+  "rt-multi-thread",
+  "macros",
+] }
+zstd = { version = "0.13.0", default-features = false }
